+++ src/pycentauri/__init__.py
@@ -15,4 +15,4 @@ __all__ = [
     "discover",
 ]
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"